Rename coordinator/http to coorindator/calls
