+++ packages/node/src/core/coordinator/index.ts
@@ -4,7 +4,7 @@ import * as apiCalls from '../requests/api-calls';
 import * as state from './state';
 import * as logger from '../logger';
 import { formatDateTime } from '../utils/date-utils';
-import * as http from './http';
+import * as calls from './calls';
 import { spawnProviderRequestProcessor } from '../providers/worker';
 
 export async function start() {
@@ -30,20 +30,20 @@ export async function start() {
   // STEP 3: Group unique API calls
   // =================================================================
   const flatApiCalls = flatMap(state2.providers, (provider) => apiCalls.flatten(provider.walletDataByIndex));
-  const aggregatedApiCalls = http.aggregate(flatApiCalls);
+  const aggregatedApiCalls = calls.aggregate(flatApiCalls);
   const state3 = state.update(state2, { aggregatedApiCalls });
   logger.logJSON('INFO', `Processing ${state3.aggregatedApiCalls.length} pending API call(s)...`);
 
   // =================================================================
   // STEP 4: Execute API calls and save the responses
   // =================================================================
-  const aggregatedCallsWithResponses = await http.callApis(state3.aggregatedApiCalls);
+  const aggregatedCallsWithResponses = await calls.callApis(state3.aggregatedApiCalls);
   const state4 = state.update(state3, { aggregatedApiCalls: aggregatedCallsWithResponses });
 
   // =================================================================
   // STEP 5: Map API responses back to each provider's API requests
   // =================================================================
-  const providersWithAPIResponses = http.disaggregate(state4);
+  const providersWithAPIResponses = calls.disaggregate(state4);
   const state5 = state.update(state4, { providers: providersWithAPIResponses });
 
   // =================================================================
